+++ doc/release-notes.rst
@@ -13,6 +13,9 @@ You can already check out a `preview <https://demo-preview.frepple.com>`_.
 
 - | Small improvements to the quick filter capability.
 
+- | The widget *Inventory history* is replaced with a widget *Inventory profile*
+    that displays both the on hand history and the projected on hand in the horizon.
+
 .. rubric:: Production planning
 
 - | The capacity report now has a row to distinguish confirmed vs proposed loading of the